requirements.txt
langchain-core
langchain-openai
streamlit
langchain 
openai 
python-dotenv
langchain-community
SpeechRecognition
pipwin
pyaudio
gTTS
google-generativeai
python-dotenv
streamlit

-e .
